feat: IFilePort uint32_t → uint64_t — removes 4GB single file limit
Changed seek/tell/size from uint32_t to uint64_t across the entire
ArcanaTS PAL interface and implementation:

- IFilePort.hpp: seek(uint64_t), tell() → uint64_t, size() → uint64_t
- ArcanaTsDb.hpp: mNextBlockOffset → uint64_t
- ArcanaTsDb.cpp: all file offset locals and constants → uint64_t
- FatFsFilePort.hpp/.cpp: matching implementation changes

Previous limit: 4GB per file (~3.4 days at 1kHz × 14B records)
New limit: exFAT max file size (theoretically 256TB)

The 3.5GB size guard remains as defense-in-depth but is no longer
the only protection against overflow.

+++ Shared/Inc/ats/ArcanaTsDb.hpp
@@ -176,7 +176,7 @@ private:
     uint8_t         mChannelCount;
     uint32_t        mNextSeqNo;
     uint32_t        mCreatedEpoch;
-    uint32_t        mNextBlockOffset;   // file offset of next data block
+    uint64_t        mNextBlockOffset;   // file offset of next data block
 
     ChannelState    mChannels[MAX_CHANNELS];
     PrimaryBuf      mPrimary;
